requirements.txt
Django>=5.0
google-genai>=1.74.0
python-dotenv>=1.2.2
